+++ app/src/main/java/com/gurumlab/vocaroutine/data/source/repository/HomeRepository.kt
@@ -1,10 +1,10 @@
-package com.gurumlab.vocaroutine.data.source.remote
+package com.gurumlab.vocaroutine.data.source.repository
 
 import com.gurumlab.vocaroutine.data.model.ApiResponse
 import com.gurumlab.vocaroutine.data.model.ListInfo
 import com.gurumlab.vocaroutine.data.model.Review
 import com.gurumlab.vocaroutine.data.source.local.AlarmDao
-import com.gurumlab.vocaroutine.data.source.repository.UserDataSource
+import com.gurumlab.vocaroutine.data.source.remote.ApiClient
 import javax.inject.Inject
 
 class HomeRepository @Inject constructor(
