+++ app/src/main/java/com/gurumlab/vocaroutine/data/source/repository/OnlineListRepository.kt
@@ -1,9 +1,10 @@
-package com.gurumlab.vocaroutine.data.source.remote
+package com.gurumlab.vocaroutine.data.source.repository
 
 import com.gurumlab.vocaroutine.data.model.ApiResponse
 import com.gurumlab.vocaroutine.data.model.ListInfo
 import com.gurumlab.vocaroutine.data.model.SharedListInfo
 import com.gurumlab.vocaroutine.data.source.local.DataStoreModule
+import com.gurumlab.vocaroutine.data.source.remote.ApiClient
 import kotlinx.coroutines.flow.first
 import javax.inject.Inject
