+++ app/src/main/java/com/gurumlab/vocaroutine/ui/list/MyListViewModel.kt
@@ -9,7 +9,7 @@ import com.gurumlab.vocaroutine.data.model.ListInfo
 import com.gurumlab.vocaroutine.data.model.onError
 import com.gurumlab.vocaroutine.data.model.onException
 import com.gurumlab.vocaroutine.data.model.onSuccess
-import com.gurumlab.vocaroutine.data.source.remote.MyListRepository
+import com.gurumlab.vocaroutine.data.source.repository.MyListRepository
 import com.gurumlab.vocaroutine.ui.common.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
