+++ app/src/main/java/com/gurumlab/vocaroutine/ui/online/OnlineListViewModel.kt
@@ -12,7 +12,7 @@ import com.gurumlab.vocaroutine.data.model.SharedListInfo
 import com.gurumlab.vocaroutine.data.model.onError
 import com.gurumlab.vocaroutine.data.model.onException
 import com.gurumlab.vocaroutine.data.model.onSuccess
-import com.gurumlab.vocaroutine.data.source.remote.OnlineListRepository
+import com.gurumlab.vocaroutine.data.source.repository.OnlineListRepository
 import com.gurumlab.vocaroutine.ui.common.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
