+++ app/src/main/java/com/gurumlab/vocaroutine/ui/setting/SettingViewModel.kt
@@ -10,7 +10,7 @@ import com.gurumlab.vocaroutine.R
 import com.gurumlab.vocaroutine.data.model.onError
 import com.gurumlab.vocaroutine.data.model.onException
 import com.gurumlab.vocaroutine.data.model.onSuccess
-import com.gurumlab.vocaroutine.data.source.remote.SettingRepository
+import com.gurumlab.vocaroutine.data.source.repository.SettingRepository
 import com.gurumlab.vocaroutine.ui.common.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
